notification/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.3
django-filter==24.1
djangorestframework==3.15.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Markdown==3.6
msgpack==1.0.8
oauthlib==3.2.2
pillow==10.3.0
pyasn1==0.6.0
psycopg2-binary==2.9.6
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-decouple==3.8
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
urllib3==2.2.1
zope.interface==6.2
